Solved user Management

+++ app/Http/Controllers/PostController.php
@@ -80,7 +80,9 @@ class PostController extends Controller
     public function edit($id)
     {
         $post =  Post::find($id);
-
+        if (auth()->user()->id != $post->user_id) {
+            return redirect("/posts")->with("error", "Unauthorized Page!");
+        }
         return view("posts.edit")->with("post", $post);
     }
 
@@ -114,6 +116,9 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (auth()->user()->id != $post->user_id) {
+            return redirect("/posts")->with("error", "Unauthorized Page!");
+        }
         $post->delete();
         return redirect("/posts")->with("success", "Post Removed");
     }
